Refactoring slack
