+++ .eslintrc
@@ -3,7 +3,8 @@
   "rules": {
     "semi": 0,
     "comma-dangle": 0,
-    "require-jsdoc": 0
+    "require-jsdoc": 0,
+    "max-len": 1
   },
   "parserOptions": {
     "ecmaVersion": 2017
